+++ packages/messaging/CHANGELOG.md
@@ -2,9 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
-## 0.1.0
+## 1.0.0
 
 ### Minor Changes
 
-- Initial package scaffold.
-
+- Add event types and payloads.
+- Add event consumers.
+- Add event dispatchers.
+- Add event repository.
+- Add event queue.